+++ choose_option.rb
@@ -1,0 +1,20 @@
+require './app'
+
+class ChooseOption
+  def choose_option(option, app)
+    case option
+    when 1
+      app.list_all_books
+    when 2
+      app.list_all_people
+    when 3
+      app.create_person
+    when 4
+      app.create_book
+    when 5
+      app.create_rental
+    when 6
+      app.list_rentals
+    end
+  end
+end
